surface skipped-file warnings on session start + recognize .har/.log as text
